CalculatePKParametersTask

R6 class for defining how pk parameters are calculated and save

Super classes

ospsuite.reportingengine::Task -> ospsuite.reportingengine::SimulationTask -> CalculatePKParametersTask

Public fields

ratioComparison

logical defining if a ratio comparison is required

Methods

Public methods

CalculatePKParametersTask$saveResults()

CalculatePKParametersTask$runTask()

CalculatePKParametersTask$clone()

Inherited methods

ospsuite.reportingengine::Task$activate()

ospsuite.reportingengine::Task$getAbsolutePath()

ospsuite.reportingengine::Task$getInputs()

ospsuite.reportingengine::Task$getRelativePath()

ospsuite.reportingengine::Task$inactivate()

ospsuite.reportingengine::Task$print()

ospsuite.reportingengine::Task$validateInput()

ospsuite.reportingengine::Task$validateStructureSetInput()

ospsuite.reportingengine::SimulationTask$initialize()

Method saveResults()

Save the task results related to a structureSet.

Usage

CalculatePKParametersTask$saveResults(structureSet, taskResults)

Arguments

structureSet

A SimulationStructure object defining the properties of a simulation set

taskResults

list of results from task run.

Method runTask()

Run task and save its output results

Usage

CalculatePKParametersTask$runTask(structureSets)

Arguments

structureSets

list of SimulationStructure objects

Method clone()

The objects of this class are cloneable with this method.

Usage

CalculatePKParametersTask$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.