ConfigurationPlan

R6 class for ConfigurationPlan guiding Qualification Workflow

Public fields

plots

list defining `PlotConfiguration` settings for specific tasks

inputs

data.frame or list mapping input files to their paths

intro

data.frame or list mapping introduction files to their paths

markdownIntro

name of markdown file that includes intro content

referenceFolder

Reference path for accessing inputs

workflowFolder

path of the output folder created or used by the Workflow.

Active bindings

sections

data.frame mapping section ids to their paths

simulationMappings

data.frame mapping simulations to their paths

observedDataSets

data.frame mapping observed datasets to their paths

Methods

Public methods

ConfigurationPlan$getSectionPath()

ConfigurationPlan$getSectionTitle()

ConfigurationPlan$getSectionMarkdown()

ConfigurationPlan$getIntroMarkdown()

ConfigurationPlan$copySectionContent()

ConfigurationPlan$copyInput()

ConfigurationPlan$copyIntro()

ConfigurationPlan$getObservedDataPath()

ConfigurationPlan$getMolWeightForObservedData()

ConfigurationPlan$getSimulationPath()

ConfigurationPlan$getPopulationPath()

ConfigurationPlan$getSimulationResultsPath()

ConfigurationPlan$getPKAnalysisResultsPath()

ConfigurationPlan$updateTheme()

Method getSectionPath()

Get location of directory corresponding to a specific section Id

Usage

ConfigurationPlan$getSectionPath(id)

Arguments

id

section identifier

Returns

The section path corresponding to the id in the configuration plan field `sections`

Method getSectionTitle()

Get markdown title to a specific section Id

Usage

ConfigurationPlan$getSectionTitle(id)

Arguments

id

section identifier

Returns

The title associated with "#" corresponding to the subection level

Method getSectionMarkdown()

Get location of .md file corresponding to a specific section Id

Usage

ConfigurationPlan$getSectionMarkdown(id)

Arguments

id

section identifier

Returns

The markdown file corresponding to the id in the configuration plan field `sections`

Method getIntroMarkdown()

Get location of .md intro file

Usage

ConfigurationPlan$getIntroMarkdown()

Returns

The markdown file corresponding to introduction

Method copySectionContent()

Copy content to section markdown

Usage

ConfigurationPlan$copySectionContent(id, logFolder = getwd())

Arguments

id

section identifier

logFolder

path where logs are saved

Method copyInput()

Copy input to section markdown

Usage

ConfigurationPlan$copyInput(input, logFolder = getwd())

Arguments

input

list including SectionId and Path

logFolder

path where logs are saved

Method copyIntro()

Copy intro to intro markdown

Usage

ConfigurationPlan$copyIntro(intro, logFolder = getwd())

Arguments

intro

list including Path

logFolder

path where logs are saved

Method getObservedDataPath()

Get location of observed data corresponding to a specific observedDataSet Id

Usage

ConfigurationPlan$getObservedDataPath(id)

Arguments

id

observedDataSet identifier

Returns

The observed data file path corresponding to the id in the configuration plan field `observedDataSet`

Method getMolWeightForObservedData()

Get molecular weight of observed data corresponding to a specific observedDataSet Id

Usage

ConfigurationPlan$getMolWeightForObservedData(id)

Arguments

id

observedDataSet identifier

Returns

The observed data file path corresponding to the id in the configuration plan field `observedDataSet`

Method getSimulationPath()

Get location of simulation file corresponding to a specific simulation and project names

Usage

ConfigurationPlan$getSimulationPath(project, simulation)

Arguments

project

name of simulation project

simulation

name of the simulation

Returns

The simulation file path corresponding to the project and simulation in the configuration plan field `simulationMappings`

Method getPopulationPath()

Get location of population file corresponding to a specific simulation and project names

Usage

ConfigurationPlan$getPopulationPath(project, simulation)

Arguments

project

name of simulation project

simulation

name of the simulation

Returns

The population file path corresponding to the project and simulation in the configuration plan field `simulationMappings`

Method getSimulationResultsPath()

Get location of simulation result file corresponding to a specific simulation and project names

Usage

ConfigurationPlan$getSimulationResultsPath(project, simulation)

Arguments

project

name of simulation project

simulation

name of the simulation

Returns

The simulation results file path corresponding to the project and simulation in the configuration plan field `simulationMappings`

Method getPKAnalysisResultsPath()

Get location of PK Analysis result file corresponding to a specific simulation and project names

Usage

ConfigurationPlan$getPKAnalysisResultsPath(project, simulation)

Arguments

project

name of simulation project

simulation

name of the simulation

Returns

The PKAnalysis results file path corresponding to the project and simulation in the configuration plan field `simulationMappings`

Method updateTheme()

Update environment theme that will be used as default during workflow

Usage

ConfigurationPlan$updateTheme()