MeanModelWorkflow

R6 class for Reporting Engine Mean Model Workflow

See also

Super class

ospsuite.reportingengine::Workflow -> MeanModelWorkflow

Public fields

simulate

`SimulationTask` object for time profile simulations

calculatePKParameters

`CalculatePKParametersTask` object for PK parameters calculation

calculateSensitivity

`SensitivityAnalysisTask` object for sensitivity analysis

plotTimeProfilesAndResiduals

`PlotTask` object for goodness of fit plots

plotMassBalance

`PlotTask` object for mass balance plot

plotAbsorption

`PlotTask` object for absorption plot

plotPKParameters

`PlotTask` object for PK parameters plot

plotSensitivity

`PlotTask` object for sensitivity plot

Methods

Public methods

MeanModelWorkflow$new()

MeanModelWorkflow$runWorkflow()

MeanModelWorkflow$clone()

Inherited methods

ospsuite.reportingengine::Workflow$activateTasks()

ospsuite.reportingengine::Workflow$getActiveTasks()

ospsuite.reportingengine::Workflow$getAllPlotTasks()

ospsuite.reportingengine::Workflow$getAllTasks()

ospsuite.reportingengine::Workflow$getInactiveTasks()

ospsuite.reportingengine::Workflow$getParameterDisplayPaths()

ospsuite.reportingengine::Workflow$getSimulationDescriptor()

ospsuite.reportingengine::Workflow$getWatermark()

ospsuite.reportingengine::Workflow$inactivateTasks()

ospsuite.reportingengine::Workflow$print()

ospsuite.reportingengine::Workflow$printReportingEngineInfo()

ospsuite.reportingengine::Workflow$setParameterDisplayPaths()

ospsuite.reportingengine::Workflow$setSimulationDescriptor()

ospsuite.reportingengine::Workflow$setWatermark()

Method new()

Create a new `MeanModelWorkflow` object.

Usage

MeanModelWorkflow$new(...)

Arguments

...

input parameters inherited from R6 class object `Workflow`.

Returns

A new `MeanModelWorkflow` object

Method runWorkflow()

Run mean model workflow tasks for all simulation sets if tasks are activated The order of tasks is as follows: # 1) Run simulations # 2) Perform PK and sensitivity analyses # 3) Perform plot tasks ## 3.a) time profiles and residual plots ## 3.b) absorption plots ## 3.c) mass balance plots ## 3.d) PK and sensitivity analyses tables and plots # 4) Render report

Usage

MeanModelWorkflow$runWorkflow()

Returns

All results and plots as a structured output in the workflow folder

Method clone()

The objects of this class are cloneable with this method.

Usage

MeanModelWorkflow$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.