QualificationWorkflow

R6 class for Reporting Engine Qualification Workflow

See also

Super class

ospsuite.reportingengine::Workflow -> QualificationWorkflow

Public fields

configurationPlan

`ConfigurationPlan` object

simulate

`SimulationTask` object for time profile simulations

calculatePKParameters

`CalculatePKParametersTask` object for PK parameters calculation

plotTimeProfiles

`PlotTask` object for time profile plots

plotComparisonTimeProfile

`PlotTask` object for comparison of time profiles plots

plotGOFMerged

`PlotTask` object for goodness of fit plots

plotPKRatio

`PlotTask` object for PK ratio plot

plotDDIRatio

`PlotTask` object for DDI ratio plot

Methods

Public methods

QualificationWorkflow$new()

QualificationWorkflow$runWorkflow()

QualificationWorkflow$updateConfigurationPlan()

QualificationWorkflow$clone()

Inherited methods

ospsuite.reportingengine::Workflow$activateTasks()

ospsuite.reportingengine::Workflow$getActiveTasks()

ospsuite.reportingengine::Workflow$getAllPlotTasks()

ospsuite.reportingengine::Workflow$getAllTasks()

ospsuite.reportingengine::Workflow$getInactiveTasks()

ospsuite.reportingengine::Workflow$getParameterDisplayPaths()

ospsuite.reportingengine::Workflow$getSimulationDescriptor()

ospsuite.reportingengine::Workflow$getWatermark()

ospsuite.reportingengine::Workflow$inactivateTasks()

ospsuite.reportingengine::Workflow$print()

ospsuite.reportingengine::Workflow$printReportingEngineInfo()

ospsuite.reportingengine::Workflow$setParameterDisplayPaths()

ospsuite.reportingengine::Workflow$setSimulationDescriptor()

ospsuite.reportingengine::Workflow$setWatermark()

Method new()

Create a new `QualificationWorkflow` object.

Usage

QualificationWorkflow$new(configurationPlan, ...)

Arguments

configurationPlan

A `ConfigurationPlan` object

...

input parameters inherited from R6 class object `Workflow`.

Returns

A new `QualificationWorkflow` object

Method runWorkflow()

Run qualification workflow tasks for all simulation sets if tasks are activated The order of tasks is as follows: # 1) Run simulations # 2) Perform PK analyses # 3) Perform plot tasks ## 3.a) time profiles and residual plots ## 3.b) comparison time profiles plots ## 3.c) PK ratio tables and plots ## 3.d) DDI ratio tables and plots # 4) Render report

Usage

QualificationWorkflow$runWorkflow()

Returns

All results and plots as a structured output in the workflow folder

Method updateConfigurationPlan()

Update the content of the workflow `configurationPlan`. Caution, updating the `configurationPlan` using this method won't update the workflow simulations and their results. Use the method only to bypass reloading a full workflow if only plot aesthetics or section content is changed.

Usage

QualificationWorkflow$updateConfigurationPlan(configurationPlanFile)

Arguments

configurationPlanFile

path to the json file corresponding to the Configuration Plan of a Qualification workflow

Method clone()

The objects of this class are cloneable with this method.

Usage

QualificationWorkflow$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.