SimulationTask

R6 class for SimulationTask settings

Super class

ospsuite.reportingengine::Task -> SimulationTask

Public fields

getTaskResults

function called by task that computes and format figure results

settings

instance of SimulationSettings class

nameTaskResults

name of function that returns task results

Methods

Public methods

SimulationTask$new()

SimulationTask$saveResults()

SimulationTask$runTask()

SimulationTask$clone()

Inherited methods

ospsuite.reportingengine::Task$activate()

ospsuite.reportingengine::Task$getAbsolutePath()

ospsuite.reportingengine::Task$getInputs()

ospsuite.reportingengine::Task$getRelativePath()

ospsuite.reportingengine::Task$inactivate()

ospsuite.reportingengine::Task$print()

ospsuite.reportingengine::Task$validateInput()

ospsuite.reportingengine::Task$validateStructureSetInput()

Method new()

Create a `SimulationTask` object

Usage

SimulationTask$new(
  getTaskResults = NULL,
  settings = NULL,
  nameTaskResults = "none",
  ...
)

Arguments

getTaskResults

function called by task that computes and format figure results

settings

instance of SimulationSettings class

nameTaskResults

name of function that returns task results

...

parameters inherited from R6 class `Task` object

Returns

A new `SimulationTask` object

Method saveResults()

Save results from task run.

Usage

SimulationTask$saveResults(set, taskResults)

Arguments

set

R6 class `SimulationStructure`

taskResults

list of results from task run.

Method runTask()

Run task and save its output

Usage

SimulationTask$runTask(structureSets)

Arguments

structureSets

list of `SimulationStructure` R6 class

Method clone()

The objects of this class are cloneable with this method.

Usage

SimulationTask$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.