TaskResults

R6 class for TaskResults settings

Public fields

id

unique identifer of results, helps in deriving final plot name

sectionId

unique identifer of section results, used only by qualification workflows

plot

`ggplot` object corresponding to the figure to be saved

plotCaption

text included into the report explaining the figure

includePlot

logical indicating if the plot should be included in final report

table

data.frame corresponding to the table to be saved

tableCaption

text included into the report explaining the table

includeTable

logical indicating if the table should be included in final report

textChunk

text included into the report explaining the table

includeTextChunk

logical indicating if the text chunk should be included in final report

Methods

Public methods

TaskResults$saveFigure()

TaskResults$saveTable()

TaskResults$addFigureToReport()

TaskResults$addTableToReport()

TaskResults$addTextChunkToReport()

Method saveFigure()

Save ggplot figure, available as plot task result, into a file

Usage

TaskResults$saveFigure(fileName, logFolder = getwd())

Arguments

fileName

path of file corresponding to the figure to save

logFolder

folder were logs are saved

Method saveTable()

Save data.frame, available as table task result, into a csv file

Usage

TaskResults$saveTable(fileName, logFolder = getwd())

Arguments

fileName

path of csv file corresponding to the table to save

logFolder

folder were logs are saved

Method addFigureToReport()

Write markdown content that adds a figure, available as plot task result, into a markdown report To be displayed, figure path must be relative to report location

Usage

TaskResults$addFigureToReport(
  reportFile,
  fileRelativePath,
  fileRootDirectory,
  logFolder = getwd()
)

Arguments

reportFile

markdown file in which the figure and its caption should be added

fileRelativePath

figure path relative to `reportFile` location

fileRootDirectory

root/working directory needed by `tracelib` package

logFolder

folder were logs are saved

Method addTableToReport()

Write markdown content that adds a data.frame, available as table task result, into a markdown report

Usage

TaskResults$addTableToReport(
  reportFile,
  fileRelativePath,
  fileRootDirectory,
  digits = NULL,
  scientific = NULL,
  logFolder = getwd()
)

Arguments

reportFile

markdown file in which the table and its caption should be added

fileRelativePath

table path relative to `reportFile` location

fileRootDirectory

root/working directory needed by `tracelib` package

digits

number of decimal digits in displayed numbers

scientific

logical defining if displayed numbers use scientific writing

logFolder

folder were logs are saved

Method addTextChunkToReport()

Write markdown content that adds text, available as textChunk task result, into a markdown report

Usage

TaskResults$addTextChunkToReport(reportFile, logFolder = getwd())

Arguments

reportFile

markdown file in which the text should be added

logFolder

folder were logs are saved