createWorkflowFromExcelInput

Creates an R script for running mean model or population workflows from an Excel file Note that the {readxl} package is required to use this function

Usage

createWorkflowFromExcelInput(
  excelFile,
  workflowFile = "workflow.R",
  removeComments = FALSE
)

Arguments

excelFile

name of the Excel file from which the R script is created

workflowFile

name of the R script file created from the Excel file

removeComments

logical to remove comments and information in workflowFile

Value

An R script of name workflowFile to be run

See also