getDefaultFolderName

Create a default folder name for reporting engine results and figures

Usage

getDefaultFolderName(..., suffix = "", sep = "-")

Arguments

...

names

suffix

Suffix to be added at the end of the folder name. Default suffix is the current date

sep

separation between names and suffix. Default is -

Value

default filename adding default suffix and file extension