inactivateWorkflowTasks

inactivates a series of `Tasks` from a `Workflow`

Usage

inactivateWorkflowTasks(workflow, tasks = workflow$getAllTasks())

Arguments

workflow

`MeanModelWorkflow` or `PopulationWorkflow` object

tasks

names of the tasks to activate Default inactivates all tasks of the workflow using workflow method `workflow$getAllTasks()`

See also

Examples

if (FALSE) { # \dontrun{

# A workflow object needs to be created first
myWorkflow <- MeanModelWorkflow$new(workflowFolder, simulationSets)

# Use enum helper to get task names
inactivateWorkflowTasks(
  workflow = myWorkflow,
  tasks = StandardSimulationTasks$simulate
)

# Default will inactivate every task of workflow
inactivateWorkflowTasks(workflow = myWorkflow)
} # }