PlotTask

R6 class for PlotTask settings

See also

Super class

ospsuite.reportingengine::Task -> PlotTask

Public fields

title

section title in the report corresponding to the task

reference

id of anchor tag referencing title

fileName

name of report appendix file associated to task

getTaskResults

function called by task that computes and format figure results

nameTaskResults

name of the function that returns task results,

Methods

Public methods

PlotTask$new()

PlotTask$saveResults()

PlotTask$runTask()

PlotTask$clone()

Inherited methods

ospsuite.reportingengine::Task$activate()

ospsuite.reportingengine::Task$getAbsolutePath()

ospsuite.reportingengine::Task$getInputs()

ospsuite.reportingengine::Task$getRelativePath()

ospsuite.reportingengine::Task$inactivate()

ospsuite.reportingengine::Task$print()

ospsuite.reportingengine::Task$validateInput()

ospsuite.reportingengine::Task$validateStructureSetInput()

Method new()

Create a PlotTask object

Usage

PlotTask$new(
  reportTitle = NULL,
  reportReference = NULL,
  fileName = NULL,
  getTaskResults = NULL,
  nameTaskResults = "none",
  ...
)

Arguments

reportTitle

title to be printed in the report

reportReference

id of anchor tag referencing title

fileName

name of report appendix file associated to task

getTaskResults

function called by task that computes and format figure results

nameTaskResults

name of the function that returns task results,

...

input parameters inherited from Task R6 class

Returns

A new PlotTask object

Method saveResults()

Save the task results related to a structureSet.

Usage

PlotTask$saveResults(structureSet, taskResults)

Arguments

structureSet

A SimulationStructure object defining the properties of a simulation set

taskResults

list of TaskResults objects

Method runTask()

Run task and save its output results

Usage

PlotTask$runTask(structureSets)

Arguments

structureSets

list of SimulationStructure objects

Method clone()

The objects of this class are cloneable with this method.

Usage

PlotTask$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.