PopulationWorkflow

R6 class for Reporting Engine Population Workflow

See also

Super class

ospsuite.reportingengine::Workflow -> PopulationWorkflow

Public fields

workflowType

Type of population workflow

simulate

`SimulationTask` object for time profile simulations

calculatePKParameters

`CalculatePKParametersTask` object for PK parameters calculation

calculateSensitivity

`SensitivityAnalysisTask` object for sensitivity analysis

plotDemography

R6 class `Task` for demography plots

plotTimeProfilesAndResiduals

`PlotTask` object for goodness of fit plots

plotPKParameters

R6 class `Task` for PK parameters plot

plotSensitivity

R6 class `Task` for sensitivity plot

Methods

Public methods

PopulationWorkflow$new()

PopulationWorkflow$runWorkflow()

PopulationWorkflow$clone()

Inherited methods

ospsuite.reportingengine::Workflow$activateTasks()

ospsuite.reportingengine::Workflow$getActiveTasks()

ospsuite.reportingengine::Workflow$getAllPlotTasks()

ospsuite.reportingengine::Workflow$getAllTasks()

ospsuite.reportingengine::Workflow$getInactiveTasks()

ospsuite.reportingengine::Workflow$getParameterDisplayPaths()

ospsuite.reportingengine::Workflow$getSimulationDescriptor()

ospsuite.reportingengine::Workflow$getWatermark()

ospsuite.reportingengine::Workflow$inactivateTasks()

ospsuite.reportingengine::Workflow$print()

ospsuite.reportingengine::Workflow$printReportingEngineInfo()

ospsuite.reportingengine::Workflow$setParameterDisplayPaths()

ospsuite.reportingengine::Workflow$setSimulationDescriptor()

ospsuite.reportingengine::Workflow$setWatermark()

Method new()

Create a new `PopulationWorkflow` object.

Usage

PopulationWorkflow$new(
  workflowType,
  simulationSets,
  workflowFolder,
  createWordReport = TRUE,
  watermark = NULL,
  simulationSetDescriptor = NULL,
  numberSections = TRUE,
  reportTitle = NULL,
  theme = NULL
)

Arguments

workflowType

Type of population workflow. Use enum `PopulationWorkflowTypes` to get list of workflow types.

simulationSets

list of `SimulationSet` R6 class objects

workflowFolder

path of the output folder created or used by the Workflow.

createWordReport

logical of option for creating Markdwon-Report only but not a Word-Report.

watermark

displayed watermark in every plot background

simulationSetDescriptor

character Descriptor of simulation sets indicated in reports

numberSections

logical defining if the report sections should be numbered

reportTitle

report title internally added as a cover page If `reportTitle` is an existing file, it will be merged to the report as cover page.

theme

A `Theme` object from `tlf` package

Returns

A new `PopulationWorkflow` object

Method runWorkflow()

Run population workflow tasks for all simulation sets if tasks are activated The order of tasks is as follows: # 1) Run simulations # 2) Perform PK and sensitivity analyses # 3) Perform plot tasks ## 3.a) time profiles and residual plots ## 3.b) demography plots ## 3.c) PK and sensitivity analyses tables and plots # 4) Render report

Usage

PopulationWorkflow$runWorkflow()

Returns

All results and plots as a structured output in the workflow folder

Method clone()

The objects of this class are cloneable with this method.

Usage

PopulationWorkflow$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.